requirement.txt
aiogram
wikipedia
deep_translator
GoogleTranslator
yt_dlp
os
qrcode
google.generativeai